requirements.txt
pyupbit
streamlit
pandas
plotly
python-dotenv
numpy
